Stop sending deleted pod events, prepare release
It was a hack to create them, requiring creating a fake v1.Pod object
and caching any k8vent annotations, and a hack on the other end to
process them.  It is the container terminated information we are
interested in, not the fact that the pod is deleted.  Processing the
latter sometimes caused the former to be lost.  Just stop faking the
pod deleted events.

Prepare 0.6.0 release.